% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_avail.r
\name{bind_avail}
\alias{bind_avail}
\title{Binds an rzmq to an available port in given range}
\usage{
bind_avail(socket, range, iface = "tcp://*", n_tries = 100)
}
\arguments{
\item{socket}{An rzmq socket object}

\item{range}{Numbers to consider (e.g. 6000:8000)}

\item{iface}{Interface to listen on}

\item{n_tries}{Number of ports to try in range}
}
\value{
The port the socket is bound to
}
\description{
Binds an rzmq to an available port in given range
}
