% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.r
\name{check_args}
\alias{check_args}
\title{Function to check arguments with which Q() is called}
\usage{
check_args(fun, iter, const = list())
}
\arguments{
\item{fun}{A function to call}

\item{iter}{Objects to be iterated in each function call}

\item{const}{A list of constant arguments passed to each function call}
}
\value{
Processed iterated argument list if 'iter' is a list
}
\description{
Function to check arguments with which Q() is called
}
\keyword{internal}
