% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.r
\name{chunk}
\alias{chunk}
\title{Subset index chunk for processing}
\usage{
chunk(x, i)
}
\arguments{
\item{x}{Index data.frame}

\item{i}{Rows to subset}
}
\value{
x[i,]
}
\description{
'attr' in `[.data.frame` takes too much CPU time
}
