% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_LSF.r
\name{cleanup}
\alias{cleanup}
\title{Will be called when exiting the `hpc` module's main loop, use to cleanup}
\usage{
cleanup()
}
\description{
Will be called when exiting the `hpc` module's main loop, use to cleanup
}

