% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{package}
\name{clustermq}
\alias{clustermq}
\alias{clustermq-package}
\title{ZeroMQ-powered cluster jobs}
\description{
Rationale This script uses rzmq to run function calls as LSF jobs. The
function supplied *MUST* be self-sufficient, i.e. load libraries and
scripts.
}
\details{
Usage
 * Q(...)     : general queuing function

Examples
 > s = function(x) x
 > Q(s, x=c(1:3), n_jobs=1)
 returns list(1,2,3)

> t = function(x) sum(x)
 > a = matrix(3:6, nrow=2)
 > Q(t, a, n_jobs=1)
 splits a by columns, sums each column, and returns list(7, 11)

TODO list
 * rerun failed jobs?
}

