% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach.r
\name{cmq_foreach}
\alias{cmq_foreach}
\title{clustermq foreach handler}
\usage{
cmq_foreach(obj, expr, envir, data)
}
\arguments{
\item{obj}{Returned from foreach::foreach, containing the following variables:
args    : Arguments passed, each as a call
argnames: character vector of arguments passed
evalenv : Environment where to evaluate the arguments
export  : character vector of variable names to export to nodes
packages: character vector of required packages
verbose : whether to print status messages [logical]
errorHandling: string of function name to call error with, e.g. "stop"}

\item{expr}{An R expression in curly braces}

\item{envir}{Environment where to evaluate the arguments}

\item{data}{Common arguments passed by register_dopcar_cmq(), e.g. n_jobs}
}
\description{
clustermq foreach handler
}
\keyword{internal}
