% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_worker_pool.r
\name{create_worker_pool}
\alias{create_worker_pool}
\title{Sets up workers for use in clustermq}
\usage{
create_worker_pool(n_jobs, data = NULL, template = list(),
  log_worker = FALSE, qsys_id = qsys_default)
}
\arguments{
\item{n_jobs}{Number of jobs to submit}

\item{data}{Set common data (function, constant args, seed)}

\item{template}{A named list of values to fill in template}

\item{log_worker}{Write a log file for each worker}

\item{qsys_id}{Character string of QSys class to use}
}
\value{
An instance of the QSys class
}
\description{
Sets up workers for use in clustermq
}
