% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{DO_CHUNK}
\alias{DO_CHUNK}
\title{Chunk of iterated arguments for the worker}
\format{An object of class \code{integer} of length 1.}
\usage{
DO_CHUNK
}
\description{
Field has to be `chunk`
}
\keyword{datasets}
