% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{DO_SETUP}
\alias{DO_SETUP}
\title{Message contains common data}
\format{An object of class \code{integer} of length 1.}
\usage{
DO_SETUP
}
\description{
This includes the function definition, common data, export
}
\keyword{datasets}
