% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onAttach}
\alias{.onAttach}
\title{Report queueing system on package attach if not set}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{default arg for compatibility}

\item{pkgname}{default arg for compatibility}
}
\description{
Report queueing system on package attach if not set
}
