% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onLoad}
\alias{.onLoad}
\title{Select the queueing system on package loading}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{default arg for compatibility}

\item{pkgname}{default arg for compatibility}
}
\description{
This is done by setting the variable 'qsys' in the package environment
to the object that contains the desired queueing system. We further call
its setup() function if it exists, and set the variable 'qsys_id' to
the scheduler we use
}
