% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_template.r
\name{fill_template}
\alias{fill_template}
\title{Fill a QSys template will values supplied}
\usage{
fill_template(template, master, values = list(), ..., log_worker = FALSE)
}
\arguments{
\item{template}{Job submission template}

\item{master}{The address of the master; this is required}

\item{values}{Values as named list}

\item{...}{Additional values as key-value pairs (take precendence)}

\item{log_worker}{Whether to log worker (default: FALSE)}
}
\value{
A filled template
}
\description{
Fill a QSys template will values supplied
}
