% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_htcondor.r
\docType{data}
\name{HTCONDOR}
\alias{HTCONDOR}
\title{HTCondor scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTCONDOR
}
\description{
Derives from QSys to provide HTCondor-specific functions
}
\keyword{internal}
