% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_lsf.r
\docType{data}
\name{LSF}
\alias{LSF}
\title{LSF scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 26.}
\usage{
LSF
}
\description{
Derives from QSys to provide LSF-specific functions
}
\keyword{datasets}
