% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_multicore.r
\docType{data}
\name{MULTICORE}
\alias{MULTICORE}
\title{Process on multiple cores on one machine}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MULTICORE
}
\description{
This makes use of rzmq messaging and sends requests via TCP/IP
}
\keyword{internal}
