% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_pbs.r
\docType{data}
\name{PBS}
\alias{PBS}
\title{PBS/Torque scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 26.}
\usage{
PBS
}
\description{
Derives from QSys to provide PBS/Torque-specific functions
}
\keyword{datasets}
