% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_LSF.r
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Objects that are linked to the package instance}
\format{An object of class \code{environment} of length 0.}
\usage{
pkg_env
}
\description{
Objects that are linked to the package instance
}
\keyword{datasets}

