% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.r
\name{Pool}
\alias{Pool}
\title{Class for basic queuing system functions}
\description{
Provides the basic functions needed to communicate between machines
This should abstract most functions of rZMQ so the scheduler
implementations can rely on the higher level functionality
}
