% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.r
\name{process_args}
\alias{process_args}
\title{Argument checks and arrange them as list item per job}
\usage{
process_args(fun, iter, const = list(), expand_grid = FALSE,
  split_array_by = NA)
}
\arguments{
\item{fun}{the function to call}

\item{const}{arguments not to vectorise over}

\item{expand_grid}{do every combination of arguments to vectorise over}

\item{split_array_by}{how to split matrices/arrays in \code{...} (default: last dimension)}

\item{export}{objects to export to computing nodes}

\item{get}{returns the result of the run (default:T)}

\item{memory}{how many Mb of memory should be reserved to run the job}

\item{seed}{random seed for the function to run}

\item{n.chunks}{how much jobs to split functions calls into (default: number of calls)}

\item{chunk.size}{how many function calls in one job (default: 1)}
}
\value{
list of job results if get=T
}
\description{
Argument checks and arrange them as list item per job
}

