% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chunk.r
\name{process_chunk}
\alias{process_chunk}
\title{Function to process a chunk of calls}
\usage{
process_chunk(df, fun, const_args, common_seed)
}
\arguments{
\item{df}{A data.frame with call IDs as rownames and arguments as columns}

\item{common_seed}{A seed offset common to all function calls}
}
\value{
A list of call results (or try-error if they failed)
}
\description{
Each chunk comes encapsulated in a data.frame
}

