% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.r
\name{proxy}
\alias{proxy}
\title{SSH proxy for different schedulers}
\usage{
proxy(master)
}
\arguments{
\item{master}{The rzmq address to connect to the master (tcp://<node>:<port>)}
}
\description{
Do not call this manually, the SSH qsys will do that
}
