% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{PROXY_CMD}
\alias{PROXY_CMD}
\title{Message is an SSH command}
\format{An object of class \code{integer} of length 1.}
\usage{
PROXY_CMD
}
\description{
Field is either `exec` (command to run) or `reply` (how it went)
}
\keyword{datasets}
