% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{PROXY_NOOP}
\alias{PROXY_NOOP}
\title{SSH proxy heartbeating}
\format{An object of class \code{integer} of length 1.}
\usage{
PROXY_NOOP
}
\description{
No fields. Answer has to be `PROXY_NOOP`
}
\keyword{datasets}
