% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{PROXY_READY}
\alias{PROXY_READY}
\title{Message ID indicating SSH proxy is ready to distribute data}
\format{An object of class \code{integer} of length 1.}
\usage{
PROXY_READY
}
\description{
Field has to be `proxy`
}
\keyword{datasets}
