% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_rettype.r
\docType{data}
\name{purrr_lookup}
\alias{purrr_lookup}
\title{Lookup table for return types to purrr functions}
\format{An object of class \code{list} of length 9.}
\usage{
purrr_lookup
}
\description{
Lookup table for return types to purrr functions
}
\keyword{datasets}
