% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach.r
\name{register_dopar_cmq}
\alias{register_dopar_cmq}
\title{Register clustermq as `foreach` parallel handler}
\usage{
register_dopar_cmq(...)
}
\arguments{
\item{...}{List of arguments passed to the `Q` function, e.g. n_jobs}
}
\description{
Register clustermq as `foreach` parallel handler
}
