% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.r
\name{run_worker}
\alias{run_worker}
\title{R worker submitted as cluster job}
\usage{
run_worker(worker_id, master, memlimit)
}
\arguments{
\item{worker_id}{The ID of the worker (usually just numbered)}

\item{master}{The master address (tcp://ip:port)}

\item{memlimit}{Maximum memory before throwing an error}
}
\description{
Do not call this manually, the master will do that
}

