% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_LSF.r
\name{send_common_data}
\alias{send_common_data}
\title{Send the data common to all workers, only serialize once}
\usage{
send_common_data(...)
}
\description{
Send the data common to all workers, only serialize once
}

