% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys.r
\docType{data}
\name{setup}
\alias{setup}
\title{Setup routines for queueing systems}
\format{An object of class \code{list} of length 1.}
\usage{
setup
}
\arguments{
\item{qsys}{A QSys-derived class}
}
\value{
A modified class
}
\description{
If there are any option(...)s handle them here
}
\keyword{datasets}

