% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_sge.r
\docType{data}
\name{SGE}
\alias{SGE}
\title{SGE scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SGE
}
\description{
Derives from QSys to provide SGE-specific functions
}
\keyword{internal}
