% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.r
\name{ssh}
\alias{ssh}
\title{SSH proxy for different schedulers}
\usage{
ssh(master_port)
}
\arguments{
\item{master}{The master address (tcp://ip:port)}
}
\description{
Do not call this manually, the SSH qsys will do that
}

