% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{SSH_NOOP}
\alias{SSH_NOOP}
\title{SSH proxy heartbeating}
\format{An object of class \code{integer} of length 1.}
\usage{
SSH_NOOP
}
\description{
No fields. Answer has to be `SSH_NOOP`
}
\keyword{datasets}
