% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_proxy.r
\name{ssh_proxy}
\alias{ssh_proxy}
\title{SSH proxy for different schedulers}
\usage{
ssh_proxy(fwd_port, qsys_id = qsys_default)
}
\arguments{
\item{qsys_id}{Character string of QSys class to use}

\item{master}{The master address to connect to (remote end of reverse tunnel)}
}
\description{
Do not call this manually, the SSH qsys will do that
}
\keyword{internal}
