% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.r
\docType{data}
\name{SSH_READY}
\alias{SSH_READY}
\title{Message ID indicating SSH proxy is ready to distribute data}
\format{An object of class \code{integer} of length 1.}
\usage{
SSH_READY
}
\description{
Field has to be `proxy`
}
\keyword{datasets}

