% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{SSH_STOP}
\alias{SSH_STOP}
\title{Message telling the SSH proxy to clean up}
\format{An object of class \code{integer} of length 1.}
\usage{
SSH_STOP
}
\description{
No fields. Signals the worker to break its main loop.
}
\keyword{datasets}

