% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{SSH_UP}
\alias{SSH_UP}
\title{Message ID indicating SSH proxy is up}
\format{An object of class \code{integer} of length 1.}
\usage{
SSH_UP
}
\description{
Message ID indicating SSH proxy is up
}
\keyword{datasets}

