% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_LSF.r
\name{submit_job}
\alias{submit_job}
\title{Submits one job to the queuing system}
\usage{
submit_job(memory, log_worker = FALSE)
}
\arguments{
\item{memory}{The amount of memory (megabytes) to request}

\item{log_worker}{Create a log file for each worker}
}
\description{
Submits one job to the queuing system
}

