% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_result.r
\name{summarize_result}
\alias{summarize_result}
\title{Print a summary of errors and warnings that occurred during processing}
\usage{
summarize_result(result, n_errors, n_warnings, cond_msgs,
  at = length(result), fail_on_error = TRUE)
}
\arguments{
\item{result}{A list or vector of the processing result}

\item{n_errors}{How many errors occurred}

\item{n_warnings}{How many warnings occurred}

\item{cond_msgs}{Error and warnings messages, we display first 50}

\item{at}{How many calls were procesed  up to this point}

\item{fail_on_error}{Stop if error(s) occurred}
}
\description{
Print a summary of errors and warnings that occurred during processing
}
\keyword{internal}
