% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_LSF.r
\docType{data}
\name{template}
\alias{template}
\title{A template string used to submit jobs}
\format{An object of class \code{character} of length 1.}
\usage{
template
}
\description{
A template string used to submit jobs
}
\keyword{datasets}

