% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel_result.r
\name{unravel_result}
\alias{unravel_result}
\title{Separate out warnings and errors from result}
\usage{
unravel_result(robj, at = length(robj$result))
}
\arguments{
\item{robj}{A list with fields 'result' and 'warnings'}

\item{at}{How many calls were procesed  up to this point}
}
\description{
Separate out warnings and errors from result
}
