% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\docType{data}
\name{vec_lookup}
\alias{vec_lookup}
\title{Lookup table for return types to vector NAs}
\format{
An object of class \code{list} of length 9.
}
\usage{
vec_lookup
}
\description{
Lookup table for return types to vector NAs
}
\keyword{internal}
