% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_chunk.r
\name{work_chunk}
\alias{work_chunk}
\title{Function to process a chunk of calls}
\usage{
work_chunk(
  df,
  fun,
  const = list(),
  rettype = "list",
  common_seed = NULL,
  progress = FALSE
)
}
\arguments{
\item{df}{A data.frame with call IDs as rownames and arguments as columns}

\item{fun}{The function to call}

\item{const}{Constant arguments passed to each call}

\item{rettype}{Return type of function}

\item{common_seed}{A seed offset common to all function calls}

\item{progress}{Logical indicated whether to display a progress bar}
}
\value{
A list of call results (or try-error if they failed)
}
\description{
Each chunk comes encapsulated in a data.frame
}
\keyword{internal}
