% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{WORKER_DONE}
\alias{WORKER_DONE}
\title{Message ID indicating worker is shutting down}
\format{An object of class \code{integer} of length 1.}
\usage{
WORKER_DONE
}
\description{
Field has to be `time` with a object returned by `Sys.time()`
}
\keyword{datasets}
