% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{WORKER_READY}
\alias{WORKER_READY}
\title{Message ID indicating worker is accepting jobs}
\format{An object of class \code{integer} of length 1.}
\usage{
WORKER_READY
}
\description{
It may contain the field `result` with a finished chunk
}
\keyword{datasets}
