% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{WORKER_STOP}
\alias{WORKER_STOP}
\title{Message ID telling worker to stop}
\format{An object of class \code{integer} of length 1.}
\usage{
WORKER_STOP
}
\description{
No fields
}
\keyword{datasets}
