% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermq-package.r
\docType{data}
\name{WORKER_UP}
\alias{WORKER_UP}
\title{Message ID indicating worker is accepting jobs}
\format{An object of class \code{integer} of length 1.}
\usage{
WORKER_UP
}
\description{
Field has to be `worker_id` to master or empty to ssh_proxy
Answer is serialized common data (`fun`, `const`, and `seed`)
or `redirect` (with URL where worker can get data)
}
\keyword{datasets}
