% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workers.r
\name{workers}
\alias{workers}
\title{Creates a pool of workers}
\usage{
workers(
  n_jobs,
  data = NULL,
  reuse = TRUE,
  template = list(),
  log_worker = FALSE,
  qsys_id = getOption("clustermq.scheduler", qsys_default),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_jobs}{Number of jobs to submit (0 implies local processing)}

\item{data}{Set common data (function, constant args, seed)}

\item{reuse}{Whether workers are reusable or get shut down after call}

\item{template}{A named list of values to fill in template}

\item{log_worker}{Write a log file for each worker}

\item{qsys_id}{Character string of QSys class to use}

\item{verbose}{Print message about worker startup}

\item{...}{Additional arguments passed to the qsys constructor}
}
\value{
An instance of the QSys class
}
\description{
Creates a pool of workers
}
