% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{fill_template}
\alias{fill_template}
\title{Fill a template string with supplied values}
\usage{
fill_template(template, values, required = c())
}
\arguments{
\item{template}{A character string of a submission template}

\item{values}{A named list of key-value pairs}

\item{required}{Keys that must be present in the template (default: none)}
}
\value{
A template where placeholder fields were replaced by values
}
\description{
Fill a template string with supplied values
}
\keyword{internal}
