% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{host}
\alias{host}
\title{Construct the ZeroMQ host}
\usage{
host(short = getOption("clustermq.short.host", TRUE))
}
\arguments{
\item{short}{whether to use unqualified host name (before first dot)}
}
\value{
the host name as character string
}
\description{
Construct the ZeroMQ host
}
