% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_htcondor.r
\docType{data}
\name{HTCONDOR}
\alias{HTCONDOR}
\title{HTCondor scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTCONDOR
}
\description{
Derives from QSys to provide HTCondor-specific functions
}
\examples{
\dontrun{
options(clustermq.scheduler="HTCONDOR")
library(clustermq)
library(foreach)
register_dopar_cmq(n_jobs=2) 
foreach(i=1:3) \%dopar\% sqrt(i) # this will be executed as jobs
}
}
\keyword{internal}
