% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_local.r
\docType{data}
\name{LOCAL}
\alias{LOCAL}
\title{Placeholder for local processing}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LOCAL
}
\description{
Mainly so tests pass without setting up a scheduler
}
\keyword{datasets}
