% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_multicore.r
\name{MULTICORE}
\alias{MULTICORE}
\title{Process on multiple cores on one machine}
\description{
Derives from QSys to provide multicore-specific functions
}
\keyword{internal}
\section{Super class}{
\code{\link[clustermq:QSys]{clustermq::QSys}} -> \code{MULTICORE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MULTICORE$new()}}
\item \href{#method-submit_jobs}{\code{MULTICORE$submit_jobs()}}
\item \href{#method-cleanup}{\code{MULTICORE$cleanup()}}
\item \href{#method-finalize}{\code{MULTICORE$finalize()}}
\item \href{#method-clone}{\code{MULTICORE$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="receive_data">}\href{../../clustermq/html/QSys.html#method-receive_data}{\code{clustermq::QSys$receive_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="send_call">}\href{../../clustermq/html/QSys.html#method-send_call}{\code{clustermq::QSys$send_call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="send_common_data">}\href{../../clustermq/html/QSys.html#method-send_common_data}{\code{clustermq::QSys$send_common_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="send_job_data">}\href{../../clustermq/html/QSys.html#method-send_job_data}{\code{clustermq::QSys$send_job_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="send_shutdown_worker">}\href{../../clustermq/html/QSys.html#method-send_shutdown_worker}{\code{clustermq::QSys$send_shutdown_worker()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="send_wait">}\href{../../clustermq/html/QSys.html#method-send_wait}{\code{clustermq::QSys$send_wait()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="clustermq" data-topic="QSys" data-id="set_common_data">}\href{../../clustermq/html/QSys.html#method-set_common_data}{\code{clustermq::QSys$set_common_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MULTICORE$new(addr = host("127.0.0.1"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit_jobs"></a>}}
\if{latex}{\out{\hypertarget{method-submit_jobs}{}}}
\subsection{Method \code{submit_jobs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MULTICORE$submit_jobs(
  n_jobs,
  ...,
  log_worker = FALSE,
  log_file = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanup"></a>}}
\if{latex}{\out{\hypertarget{method-cleanup}{}}}
\subsection{Method \code{cleanup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MULTICORE$cleanup(quiet = FALSE, timeout = 3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MULTICORE$finalize(quiet = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MULTICORE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
