% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.r
\name{proxy}
\alias{proxy}
\title{SSH proxy for different schedulers}
\usage{
proxy(ctl, job)
}
\arguments{
\item{ctl}{The port to connect to the master for proxy control}

\item{job}{The port to connect to the master for job control}
}
\description{
Do not call this manually, the SSH qsys will do that
}
