% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys.r
\docType{data}
\name{QSys}
\alias{QSys}
\title{Class for basic queuing system functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QSys
}
\description{
Provides the basic functions needed to communicate between machines
This should abstract most functions of rZMQ so the scheduler
implementations can rely on the higher level functionality
}
\keyword{datasets}
