% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_slurm.r
\docType{data}
\name{SLURM}
\alias{SLURM}
\title{SLURM scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SLURM
}
\description{
Derives from QSys to provide SLURM-specific functions
}
\keyword{datasets}
