% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_ssh.r
\docType{data}
\name{SSH}
\alias{SSH}
\title{SSH scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SSH
}
\description{
Derives from QSys to provide SSH-specific functions
}
\keyword{internal}
