% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsys_ssh.r
\docType{data}
\name{SSH}
\alias{SSH}
\title{LSF scheduler functions}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SSH
}
\description{
Derives from QSys to provide LSF-specific functions
}
\keyword{datasets}
