% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_proxy.r
\name{ssh_proxy}
\alias{ssh_proxy}
\title{SSH proxy for different schedulers}
\usage{
ssh_proxy(master_port)
}
\arguments{
\item{master_port}{The port for SSH reverse tunnel to master}
}
\description{
Do not call this manually, the SSH qsys will do that
}
