% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_proxy.r
\name{ssh_proxy}
\alias{ssh_proxy}
\title{SSH proxy for different schedulers}
\usage{
ssh_proxy(ctl, job, qsys_id = qsys_default)
}
\arguments{
\item{ctl}{The port to connect to the master for proxy control}

\item{job}{The port to connect to the master for job control}

\item{qsys_id}{Character string of QSys class to use}
}
\description{
Do not call this manually, the SSH qsys will do that
}
\keyword{internal}
