% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.r
\name{worker}
\alias{worker}
\title{R worker submitted as cluster job}
\usage{
worker(master, ..., verbose = TRUE, context = NULL)
}
\arguments{
\item{master}{The master address (tcp://ip:port)}

\item{...}{Catch-all to not break older template values (ignored)}

\item{verbose}{Whether to print debug messages}

\item{context}{ZeroMQ context (for internal testing)}
}
\description{
Do not call this manually, the master will do that
}
\keyword{internal}
