% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.r
\name{worker}
\alias{worker}
\title{R worker submitted as cluster job}
\usage{
worker(master, timeout = 600, ..., verbose = TRUE)
}
\arguments{
\item{master}{The master address (tcp://ip:port)}

\item{timeout}{Time until worker shuts down without hearing from master}

\item{...}{Catch-all to not break older template values (ignored)}

\item{verbose}{Whether to print debug messages}
}
\description{
Do not call this manually, the master will do that
}
