% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeromq.r
\name{ZeroMQ}
\alias{ZeroMQ}
\title{Wrap C++ Rcpp module in R6 to get reliable argument matching}
\description{
This is an R6 wrapper of the C++ class in order to support R argument
matching. Ideally, Rcpp will at some point support this natively and this
file will no longer be necessary. Until then, it causes redundancy with
zeromq.cpp, but this is a small inconvenience and much less error-prone than
only relying on positional arguments.
}
